Remove 'export' from LD_LIBRARY_PATH.
This sets LD_LIBRARY_PATH for the `java` executable but _not_ to any
subprocesses launched by Java. This should eliminate any unintentional
side effects of using this wrapper.

It should be noted that LD_LIBRARY_PATH is also not set on any Java
subprocesses if they are launched using a direct path to the underlying
`java` executable. It is generally not possible to predict whether this
is desired or not.